container_engine/django_tutorial/requirements.txt
Django==1.11.10
mysqlclient==1.3.12
wheel==0.30.0
gunicorn==19.7.1
psycopg2==2.7.4
google.cloud.logging
google-python-cloud-debugger
google-auth
